refactor(api)!: align final package ownership
