+++ src/main/java/io/github/ggomarighetti/jparsqlsearch/autoconfigure/JpaRsqlSearchAutoConfiguration.java
@@ -4,7 +4,7 @@ import io.github.ggomarighetti.jparsqlsearch.definition.SearchDefinitionFactory;
 import io.github.ggomarighetti.jparsqlsearch.compile.SearchCompiler;
 import io.github.ggomarighetti.jparsqlsearch.jpa.JpaSearchDefinitionValidator;
 import io.github.ggomarighetti.jparsqlsearch.rsql.engine.SearchRsqlEngine;
-import io.github.ggomarighetti.jparsqlsearch.validation.SearchDefinitionValidator;
+import io.github.ggomarighetti.jparsqlsearch.definition.validation.SearchDefinitionValidator;
 import jakarta.persistence.EntityManagerFactory;
 import org.springframework.beans.factory.ObjectProvider;
 import org.springframework.boot.autoconfigure.AutoConfiguration;
