+++ src/test/java/io/github/ggomarighetti/jparsqlsearch/compile/RsqlPropertyTest.java
@@ -8,7 +8,7 @@ import static org.junit.jupiter.api.Assertions.fail;
 
 import io.github.ggomarighetti.jparsqlsearch.definition.SearchDefinition;
 import io.github.ggomarighetti.jparsqlsearch.exception.RsqlFilterValidationException;
-import io.github.ggomarighetti.jparsqlsearch.exception.SearchProtectionException;
+import io.github.ggomarighetti.jparsqlsearch.protection.SearchProtectionException;
 import io.github.ggomarighetti.jparsqlsearch.integration.bench.domain.Product;
 import io.github.ggomarighetti.jparsqlsearch.rsql.RsqlCompilationRequest;
 import io.github.ggomarighetti.jparsqlsearch.property.RsqlInputGenerator;
