+++ src/test/java/io/github/ggomarighetti/jparsqlsearch/unit/SearchCompilerTest.java
@@ -2,7 +2,7 @@ package io.github.ggomarighetti.jparsqlsearch.unit;
 
 import io.github.ggomarighetti.jparsqlsearch.rsql.backend.perplexhub.PerplexhubRsqlEngines;
 import io.github.ggomarighetti.jparsqlsearch.definition.SearchDefinition;
-import io.github.ggomarighetti.jparsqlsearch.exception.SearchProtectionException;
+import io.github.ggomarighetti.jparsqlsearch.protection.SearchProtectionException;
 import io.github.ggomarighetti.jparsqlsearch.compile.CompiledSearch;
 import io.github.ggomarighetti.jparsqlsearch.compile.SearchCompiler;
 import io.github.ggomarighetti.jparsqlsearch.policy.SearchPolicy;
